Module version checking
Emit symbols in the binary which let the linker issue an error if the swiftmodule file does not match the generated archive/dylib.

Store an MD5 hash in the swiftmodule file and generate a symbol in the binary which encodes that hash.
For each imported module, reference the module hash symbol.
If the swiftmodule does not match, the linker will issue an undefined-symbol error.

Of course, this is only done for modules which are not compiled with library evolution.

In the current implementation version checking only works with whole-module-compilation.