Fix use of Apple target macros
For some reason, `TARGET_OS_IOS` is defined to be 1 on iOS and defined to 0 on
all other Apple platforms.  So you can't just check whether the symbol is
defined or not.

It probably would suffice to just use the value, but I had headaches a while
back with some compiler that insisted on complaining about undefined
preprocessor macros, so I've used the horribly conservative `defined(FOO) &&
FOO` construction here.