[ObjC] Only Diagnose Selector Conflicts From Classes
Clang 1200 appears to be returning us results from protocol
declarations when we run lookup for an Objective-C selector. This code
path can therefore see protocol requirements, and would erroneously
diagnose witnesses from Swift subclasses of Objective-C classes
as an attempt to conflict with the requirement in the
protocol itself.

This fix is super narrow to try to unblock our builds. Tests will follow
once I can reduce the issue.

Resolves rdar://60005962