[ownership] Now that cond_br only takes trivial operands in ossa, eliminate all special code for handling it.
I banned this in an earlier commit since it wasn't really useful and introduced
a bunch of unneeded complexity into the ownership system (namely the name for
BranchPropagatedUser). Now that we are free of that, we can clean up the
ownership code and just treat cond_br simply! I left in a special section in
OperandOwnership so I could put in a nice comment there.